@@ -1,10 +1,11 @@
 package com.xkcoding.orm.jpa.entity;
 
 import com.xkcoding.orm.jpa.entity.base.AbstractAuditModel;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
 import lombok.*;
 
-import javax.persistence.*;
-import java.util.Collection;
 import java.util.Date;
 
 /**
@@ -15,14 +16,13 @@ import java.util.Date;
  * @author yangkai.shen
  * @date Created in 2018-11-07 14:06
  */
-@EqualsAndHashCode(callSuper = true)
-@NoArgsConstructor
-@AllArgsConstructor
 @Data
 @Builder
 @Entity
 @Table(name = "orm_user")
-@ToString(callSuper = true)
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode(callSuper = true)
 public class User extends AbstractAuditModel {
     /**
      * 用户名
@@ -60,20 +60,4 @@ public class User extends AbstractAuditModel {
      */
     @Column(name = "last_login_time")
     private Date lastLoginTime;
-
-    /**
-     * 关联部门表
-     * 1、关系维护端，负责多对多关系的绑定和解除
-     * 2、@JoinTable注解的name属性指定关联表的名字，joinColumns指定外键的名字，关联到关系维护端(User)
-     * 3、inverseJoinColumns指定外键的名字，要关联的关系被维护端(Department)
-     * 4、其实可以不使用@JoinTable注解，默认生成的关联表名称为主表表名+下划线+从表表名，
-     * 即表名为user_department
-     * 关联到主表的外键名：主表名+下划线+主表中的主键列名,即user_id,这里使用referencedColumnName指定
-     * 关联到从表的外键名：主表中用于关联的属性名+下划线+从表的主键列名,department_id
-     * 主表就是关系维护端对应的表，从表就是关系被维护端对应的表
-     */
-    @ManyToMany(cascade = CascadeType.ALL, fetch = FetchType.EAGER)
-    @JoinTable(name = "orm_user_dept", joinColumns = @JoinColumn(name = "user_id", referencedColumnName = "id"), inverseJoinColumns = @JoinColumn(name = "dept_id", referencedColumnName = "id"))
-    private Collection<Department> departmentList;
-
 }
