@@ -1,11 +1,11 @@
 package com.xkcoding.orm.jpa.entity.base;
 
+import jakarta.persistence.*;
 import lombok.Data;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import javax.persistence.*;
 import java.io.Serializable;
 import java.util.Date;
 
@@ -17,9 +17,9 @@ import java.util.Date;
  * @author yangkai.shen
  * @date Created in 2018-11-07 14:01
  */
+@Data
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-@Data
 public abstract class AbstractAuditModel implements Serializable {
     /**
      * 主键
